+++ Tests/UIKitTests/UINavigationControllerExtensionsTests.swift
@@ -53,5 +53,19 @@ final class UINavigationControllerExtensionsTests: XCTestCase {
         XCTAssertNotNil(color)
         XCTAssertEqual(color!, .red)
     }
+    
+#if !os(tvOS)
+    func testHideBottomBar() {
+        let rootVC = UIViewController()
+        let navigationController = UINavigationController(rootViewController: rootVC)
+        let vcToPush = UIViewController()
+        let tabVC = UITabBarController()
+        tabVC.viewControllers = [navigationController]
+        navigationController.pushViewController(vcToPush, hidesBottomBar: true, animated: false)
+        XCTAssert(vcToPush.hidesBottomBarWhenPushed)
+        XCTAssert(tabVC.tabBar.isHidden)
+    }
+#endif
+    
 }
 #endif